requirements.txt
psutil>=5.9
requests>=2.32
rich>=13.7
python-nmap>=0.7.1
scapy>=2.5


